Okay, apparently sometimes OBA will return an ArrivalAndDeparture object that says it's predicted, but the predictedArrivalTime is 0.  This is why the page will sometimes say a bus is a thousand minutes early, and probably accounts for some of the bugginess I've seen in the svgs.
It looks like the OBA page just considers those "not predicted" so I'm updating scripts to do the same thing.  Now everything should be checking predictedArrivalTime to make sure it's not 0 before we believe it's predicted.

(Also adjusting font size on the page, because I don't feel like breaking that out into a separate change.)